+++ CHANGELOG.md
@@ -18,6 +18,7 @@ CHANGELOG
 
 #### Parsers
 - `intelmq.bots.parsers.misp`: Fix Object attribute (#1318).
+- `intelmq.bots.parsers.cymru.parser_cap_program`: Add support for new format (extra data about botnet of 'bots').
 
 #### Experts